+++ src/main/java/net/geforcemods/securitycraft/blocks/BlockFakeWater.java
@@ -53,7 +53,7 @@ public class BlockFakeWater extends BlockDynamicLiquid{
 	@Override
 	public void updateTick(World world, BlockPos pos, IBlockState state, Random rand)
 	{
-		int level = state.getValue(LEVEL).intValue();
+		int level = state.getValue(LEVEL);
 		byte levelToAdd = 1;
 
 		if (this.material == Material.LAVA && !world.provider.doesWaterVaporize())
@@ -92,7 +92,7 @@ public class BlockFakeWater extends BlockDynamicLiquid{
 
 				if (stateBelow.getMaterial().isSolid())
 					nextLevel = 0;
-				else if (stateBelow.getMaterial() == this.material && stateBelow.getValue(LEVEL).intValue() == 0)
+				else if (stateBelow.getMaterial() == this.material && stateBelow.getValue(LEVEL) == 0)
 					nextLevel = 0;
 			}
 
@@ -184,7 +184,7 @@ public class BlockFakeWater extends BlockDynamicLiquid{
 				BlockPos offsetPos = pos.offset(facing);
 				IBlockState offsetState = world.getBlockState(offsetPos);
 
-				if (!isBlocked(world, offsetPos, offsetState) && (offsetState.getBlock().getMaterial(offsetState) != this.material || offsetState.getValue(LEVEL).intValue() > 0))
+				if (!isBlocked(world, offsetPos, offsetState) && (offsetState.getBlock().getMaterial(offsetState) != this.material || offsetState.getValue(LEVEL) > 0))
 				{
 					if (!isBlocked(world, offsetPos.down(), offsetState))
 						return distance;
@@ -215,7 +215,7 @@ public class BlockFakeWater extends BlockDynamicLiquid{
 			BlockPos offsetPos = pos.offset(facing);
 			IBlockState offsetState = world.getBlockState(offsetPos);
 
-			if (!isBlocked(world, offsetPos, offsetState) && (offsetState.getBlock().getMaterial(offsetState) != this.material || offsetState.getValue(LEVEL).intValue() > 0))
+			if (!isBlocked(world, offsetPos, offsetState) && (offsetState.getBlock().getMaterial(offsetState) != this.material || offsetState.getValue(LEVEL) > 0))
 			{
 				int oppositeCost;
